pyproject.toml
[project]
name = "directus_git_sync"
description = "import/export from directus"
version = "0.0.1"
authors = [{ name = "Bea Steers", email = "bea.steers@gmail.com" }]
dependencies = ["requests", "PyYAML", "starstar"]

[project.scripts]
directus-git-sync = "directus_git_sync:cli"